adding crud

+++ src/main/java/org/example/tahadaw/DTO/IN/GiftHistoryCreateDTOIn.java
@@ -23,7 +23,7 @@ public class GiftHistoryCreateDTOIn {
 
     private String occasionType;
     private LocalDate giftDate;
-    private Long priceMinor;
+    private Double priceMinor;
     private Boolean wasGifted;
     private Integer userRating;
     private String notes;
